use 10 threads

+++ experiments/thread_pool/test/task_test.h
@@ -27,7 +27,7 @@ public:
     }
 
 private:
-    int num__;
+    unsigned long num__;
     std::string name__;
 };
